+++ src/resources/addressing/address-maps/accounts.ts
@@ -1,6 +1,7 @@
 // File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
 
 import { APIResource } from '../../../resource';
+import { isRequestOptions } from '../../../core';
 import * as Core from '../../../core';
 
 export class Accounts extends APIResource {
@@ -24,7 +25,7 @@ export class Accounts extends APIResource {
     params: AccountUpdateParams,
     options?: Core.RequestOptions,
   ): Core.APIPromise<AccountUpdateResponse> {
-    const { account_id, body } = params;
+    const { account_id = this._client.accountId, body } = params;
     return this._client.put(
       `/accounts/${account_id}/addressing/address_maps/${addressMapId}/accounts/${account_id}`,
       { body: body, ...options },
@@ -45,10 +46,19 @@ export class Accounts extends APIResource {
    */
   delete(
     addressMapId: string,
-    params: AccountDeleteParams,
+    params?: AccountDeleteParams,
+    options?: Core.RequestOptions,
+  ): Core.APIPromise<AccountDeleteResponse>;
+  delete(addressMapId: string, options?: Core.RequestOptions): Core.APIPromise<AccountDeleteResponse>;
+  delete(
+    addressMapId: string,
+    params: AccountDeleteParams | Core.RequestOptions = {},
     options?: Core.RequestOptions,
   ): Core.APIPromise<AccountDeleteResponse> {
-    const { account_id } = params;
+    if (isRequestOptions(params)) {
+      return this.delete(addressMapId, {}, params);
+    }
+    const { account_id = this._client.accountId } = params;
     return this._client.delete(
       `/accounts/${account_id}/addressing/address_maps/${addressMapId}/accounts/${account_id}`,
       options,
@@ -122,6 +132,11 @@ export namespace AccountUpdateResponse {
      * Total results available without any search parameters.
      */
     total_count?: number;
+
+    /**
+     * The number of total pages in the entire result set.
+     */
+    total_pages?: number;
   }
 }
 
@@ -191,6 +206,11 @@ export namespace AccountDeleteResponse {
      * Total results available without any search parameters.
      */
     total_count?: number;
+
+    /**
+     * The number of total pages in the entire result set.
+     */
+    total_pages?: number;
   }
 }
 
@@ -198,10 +218,10 @@ export interface AccountUpdateParams {
   /**
    * Path param: Identifier of a Cloudflare account.
    */
-  account_id: string;
+  account_id?: string;
 
   /**
-   * Body param:
+   * Body param
    */
   body: unknown;
 }
@@ -210,7 +230,7 @@ export interface AccountDeleteParams {
   /**
    * Identifier of a Cloudflare account.
    */
-  account_id: string;
+  account_id?: string;
 }
 
 export declare namespace Accounts {
