+++ src/resources/addressing/address-maps/zones.ts
@@ -1,6 +1,7 @@
 // File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
 
 import { APIResource } from '../../../resource';
+import { isRequestOptions } from '../../../core';
 import * as Core from '../../../core';
 
 export class Zones extends APIResource {
@@ -25,7 +26,7 @@ export class Zones extends APIResource {
     params: ZoneUpdateParams,
     options?: Core.RequestOptions,
   ): Core.APIPromise<ZoneUpdateResponse> {
-    const { zone_id, account_id, body } = params;
+    const { zone_id = this._client.zoneId, account_id = this._client.accountId, body } = params;
     return this._client.put(
       `/accounts/${account_id}/addressing/address_maps/${addressMapId}/zones/${zone_id}`,
       { body: body, ...options },
@@ -49,10 +50,19 @@ export class Zones extends APIResource {
    */
   delete(
     addressMapId: string,
-    params: ZoneDeleteParams,
+    params?: ZoneDeleteParams,
+    options?: Core.RequestOptions,
+  ): Core.APIPromise<ZoneDeleteResponse>;
+  delete(addressMapId: string, options?: Core.RequestOptions): Core.APIPromise<ZoneDeleteResponse>;
+  delete(
+    addressMapId: string,
+    params: ZoneDeleteParams | Core.RequestOptions = {},
     options?: Core.RequestOptions,
   ): Core.APIPromise<ZoneDeleteResponse> {
-    const { zone_id, account_id } = params;
+    if (isRequestOptions(params)) {
+      return this.delete(addressMapId, {}, params);
+    }
+    const { zone_id = this._client.zoneId, account_id = this._client.accountId } = params;
     return this._client.delete(
       `/accounts/${account_id}/addressing/address_maps/${addressMapId}/zones/${zone_id}`,
       options,
@@ -126,6 +136,11 @@ export namespace ZoneUpdateResponse {
      * Total results available without any search parameters.
      */
     total_count?: number;
+
+    /**
+     * The number of total pages in the entire result set.
+     */
+    total_pages?: number;
   }
 }
 
@@ -195,6 +210,11 @@ export namespace ZoneDeleteResponse {
      * Total results available without any search parameters.
      */
     total_count?: number;
+
+    /**
+     * The number of total pages in the entire result set.
+     */
+    total_pages?: number;
   }
 }
 
@@ -202,15 +222,15 @@ export interface ZoneUpdateParams {
   /**
    * Path param: Identifier of a zone.
    */
-  zone_id: string;
+  zone_id?: string;
 
   /**
    * Path param: Identifier of a Cloudflare account.
    */
-  account_id: string;
+  account_id?: string;
 
   /**
-   * Body param:
+   * Body param
    */
   body: unknown;
 }
@@ -219,12 +239,12 @@ export interface ZoneDeleteParams {
   /**
    * Identifier of a zone.
    */
-  zone_id: string;
+  zone_id?: string;
 
   /**
    * Identifier of a Cloudflare account.
    */
-  account_id: string;
+  account_id?: string;
 }
 
 export declare namespace Zones {
